refactor: 메인페이지 컴포넌트/훅 분리, 글로벌 폰트 SUITE 적용, 이벤트 유형 미노출
## 변경 사항

### 1. 글로벌 폰트 변경 (SUITE)
- Pretendard → SUITE(Sweet) 폰트로 전환 (300~900 웨이트 대응)
- globals.css에 @font-face 선언, variables.css의 --font-sans 변수 업데이트

### 2. 메인 페이지 리팩토링 (아키텍처 규칙 적용)
- useHome.ts: 모든 상태/이펙트/fetch 로직 커스텀 훅으로 추출 (규칙 3)
- _components/HeroSection/: 히어로 무한 롤링 UI 독립 컴포넌트 (규칙 4, 5)
- _components/EventListSection/: 필터/검색/카드/페이지네이션 독립 컴포넌트 (규칙 4, 5)
- page.tsx: 16줄의 얇은 래퍼로 축소 (규칙 3)
- 각 컴포넌트별 독립 CSS 모듈 적용 (규칙 2)

### 3. 히어로 섹션 개선
- 롤링 카드 높이를 콘텐츠에 맞게 유연하게 변경 (height: auto)
- 헤더 높이(60px) 감안하여 뷰포트에 딱 맞게 조정

### 4. 이벤트 상세 페이지
- 이벤트 유형(Type) 필드 미노출 처리
