+++ frontend/src/app/_components/EventListSection/EventListSection.module.css
@@ -1,0 +1,75 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  gap: var(--space-48);
+  padding: var(--space-48) var(--space-24);
+  width: 100%;
+  max-width: 1760px;
+  margin: 0 auto;
+}
+
+.filterSection {
+  display: flex;
+  flex-direction: column;
+  gap: var(--space-16);
+  width: 100%;
+}
+
+.filterBar {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+  gap: var(--space-16);
+}
+
+.searchBox {
+  width: 400px;
+  flex-shrink: 0;
+}
+
+.dropdownGroup {
+  display: flex;
+  flex-direction: row;
+  gap: var(--space-8);
+  flex-wrap: wrap;
+  justify-content: flex-end;
+}
+
+.activeFilters {
+  display: flex;
+  flex-wrap: wrap;
+  gap: var(--space-8);
+  min-height: 38px;
+}
+
+.content {
+  width: 100%;
+  padding-bottom: var(--space-48);
+}
+
+.paginationContainer {
+  display: flex;
+  justify-content: center;
+  margin-top: var(--space-48);
+}
+
+/* 반응형 모바일 */
+@media screen and (max-width: 1024px) {
+  .filterBar {
+    flex-direction: column;
+    align-items: flex-start;
+  }
+
+  .dropdownGroup {
+    width: 100%;
+    overflow-x: auto;
+    flex-wrap: nowrap;
+    justify-content: flex-start;
+  }
+
+  .searchBox {
+    width: 100%;
+  }
+}
