+++ frontend/src/app/_components/EventListSection/index.ts
@@ -1,0 +1,1 @@
+export { default } from './EventListSection';
