+++ frontend/src/app/_components/HeroSection/index.ts
@@ -1,0 +1,1 @@
+export { default } from './HeroSection';
