--- frontend/src/app/page.module.css
@@ -1,146 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  gap: var(--space-48);
-  padding: var(--space-48) var(--space-24);
-  width: 100%;
-  max-width: 1760px;
-  margin: 0 auto;
-}
-
-.heroSection {
-  width: 100%;
-  background-color: var(--color-surface);
-  /* 회색 배경 가로로 꽉 차게 */
-  padding: 100px 0;
-  /* 위아래 여백을 아주 충분히 줌 */
-  display: flex;
-  justify-content: center;
-}
-
-.heroInner {
-  width: 100%;
-  max-width: 1278px;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  gap: 80px;
-  /* 텍스트 영역과 오른쪽 이미지 박스 사이 공간 더 충분히 넓힘 */
-  padding: 0 var(--space-24);
-}
-
-.heroContent {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  gap: var(--space-16);
-  max-width: 50%;
-}
-
-.heroTitle {
-  font-size: 40px;
-  font-weight: 700;
-  line-height: 1.2;
-  color: var(--color-text-gray-900);
-}
-
-.heroSubtitle {
-  font-size: 16px;
-  line-height: 1.6;
-  color: var(--color-text-gray-500);
-}
-
-.heroImageBox {
-  width: 440px;
-  height: 300px;
-  background-color: #D1D5DB;
-  /* placeholder color */
-  border-radius: var(--radius-xl);
-}
-
-.filterSection {
-  display: flex;
-  flex-direction: column;
-  gap: var(--space-16);
-  width: 100%;
-}
-
-.filterBar {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  gap: var(--space-16);
-}
-
-.searchBox {
-  width: 400px;
-  flex-shrink: 0;
-}
-
-.dropdownGroup {
-  display: flex;
-  flex-direction: row;
-  gap: var(--space-8);
-  flex-wrap: wrap;
-  justify-content: flex-end;
-}
-
-.activeFilters {
-  display: flex;
-  flex-wrap: wrap;
-  gap: var(--space-8);
-  min-height: 38px;
-}
-
-.content {
-  width: 100%;
-  padding-bottom: var(--space-48);
-}
-
-.paginationContainer {
-  display: flex;
-  justify-content: center;
-  margin-top: var(--space-48);
-}
-
-/* 반응형 모바일 */
-@media screen and (max-width: 1024px) {
-  .heroSection {
-    padding: var(--space-32) 0;
-  }
-
-  .heroInner {
-    flex-direction: column;
-    text-align: center;
-  }
-
-  .heroContent {
-    max-width: 100%;
-    align-items: center;
-  }
-
-  .heroImageBox {
-    width: 100%;
-    max-width: 600px;
-    height: 250px;
-  }
-
-  .filterBar {
-    flex-direction: column;
-    align-items: flex-start;
-  }
-
-  .dropdownGroup {
-    width: 100%;
-    overflow-x: auto;
-    flex-wrap: nowrap;
-    justify-content: flex-start;
-  }
-
-  .searchBox {
-    width: 100%;
-  }
-}
